FIX: Fixed failing tests due to deserialization
The TranslatedNonMatchingResponseBody is required by the VSP for deserialization
however when testing it is required to be able to serialise this class. The solution
was to create a test class that inherits TranslatedNonMatchingResponseBody which has
the ability to perform serialisation and use this instead for tests.

Added a TestTranslatedNonMatchingResponseBody
