+++ src/main/java/uk/gov/ida/verifyserviceprovider/dto/TestTranslatedNonMatchingResponseBody.java
@@ -1,0 +1,36 @@
+package uk.gov.ida.verifyserviceprovider.dto;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.Optional;
+
+public class TestTranslatedNonMatchingResponseBody extends TranslatedNonMatchingResponseBody {
+
+    @JsonCreator
+    public TestTranslatedNonMatchingResponseBody(
+            @JsonProperty("scenario") NonMatchingScenario scenario,
+            @JsonProperty("pid") String pid,
+            @JsonProperty("levelOfAssurance") LevelOfAssurance levelOfAssurance,
+            @JsonProperty("attributes") NonMatchingAttributes attributes
+    ) {
+        super(scenario, pid, levelOfAssurance, attributes);
+    }
+
+    public NonMatchingScenario getScenario() {
+        return super.getScenario();
+    }
+
+    public String getPid() {
+        return pid;
+    }
+
+    public LevelOfAssurance getLevelOfAssurance() {
+        return levelOfAssurance;
+    }
+
+    public Optional<NonMatchingAttributes> getAttributes() {
+        return Optional.ofNullable(attributes);
+    }
+
+}
